+++ Gemfile
@@ -48,4 +48,5 @@ group :test do
   gem 'simplecov', '~> 0.10', '< 0.18', require: false
   gem 'webdrivers'
   gem 'webmock'
+  gem 'database_cleaner'
 end
