+++ app/controllers/offers_controller.rb
@@ -84,8 +84,8 @@ class OffersController < ApplicationController
   end
 
   def create_predicate
-    # captcha_predicate = !@captcha_required || verify_recaptcha(model: @offer)
-    captcha_predicate = true
+    captcha_predicate = !@captcha_required || verify_recaptcha(model: @offer)
+    # captcha_predicate = true
     captcha_predicate && @offer.save && @offer.reload
   end
