+++ app/jobs/check_linkpay_status_job.rb
@@ -3,6 +3,8 @@ class CheckLinkpayStatusJob < ApplicationJob
 
   def perform(payment_order_id)
     payment_order = PaymentOrder.find(payment_order_id)
+    p "++++++++"
+    p payment_order_id
     return unless payment_order_valid?(payment_order)
 
     payment_order.check_linkpay_status
